+++ android/app/src/main/kotlin/bar/bto/gwa/GWAApp.kt
@@ -1,0 +1,13 @@
+package bar.bto.gwa
+
+import android.app.Application
+import timber.log.Timber
+
+class GWAApp : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        if (BuildConfig.DEBUG) {
+            Timber.plant(Timber.DebugTree())
+        }
+    }
+}